inline_color: Removed the trimming of leading whitespace
Trimming leading whitespace was incorrect, so it has been removed. Also,
reordered the filters used in the page layout so that inline_color is
called before replace. This was necessary because nokogiri breaks some
lines and thus the output of the inline_color filter would have line
breaks that were not appropriately indented for use in the yaml file
resulting from the page layout.