layouts: updated assignment layout
Updated assignment layout so that it does not include keys that have no
value, thus overwriting values that existed on the server. The user can
provide a value for a key which would then update the server, but the
default will be to not change server values for missing keys.