tags: Updated assignment to allow for user input
The assignment tag now accepts a parameter indicating the name of the
assignment whose id is being looked up. In case there is no parameter
given, then the tag will lookup the id as in the past, i.e., according
to the title of the page where the tag is being used.

This modification allows the tag to be used to generate links to
assignments instead of hard-coding assignment ids into links as was done
previously.