filters: Added liquify filter
Added liquify filter so that liquid code can be used inside of the YAML
front-matter. This is of interest for the upcoming schedule layout where
meeting information can be specified via YAML. In this case, it is
sometimes necessary to include liquid code, for example, to get an
assignment link.